% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentiles.R
\name{calculate_percentiles}
\alias{calculate_percentiles}
\title{Calculate percentiles}
\usage{
calculate_percentiles(data, value_col)
}
\arguments{
\item{data}{A data.frame}

\item{value_col}{character name of column containing values}
}
\value{
A vector of numerics with percentile values of length of nrow(data)
}
\description{
Calculate percentiles for values in a data.frame
}
\author{
Peter Marquardt
}
